further message cleanup
